retrieve cumulative activities from frontend
