+++ app/scenes/ActivitySummary/index.js
@@ -28,7 +28,7 @@ import { parseAppletEvents } from "../../models/json-ld";
 import BaseText from "../../components/base_text/base_text";
 import { newAppletSelector } from "../../state/app/app.selectors";
 import { setActivities, setCumulativeActivities } from "../../state/activities/activities.actions";
-import { getScoreFromResponse, evaluateScore, getMaxScore } from "../../services/scoring";
+import { evaluateCumulatives } from "../../services/scoring";
 
 let markdownItInstance = MarkdownIt({ typographer: true })
   .use(markdownContainer)
@@ -96,83 +96,7 @@ const ActivitySummary = ({ responses, activity, applet, cumulativeActivities, se
   const [messages, setMessages] = useState([]);
 
   useEffect(() => {
-    const parser = new Parser({
-      logical: true,
-      comparison: true,
-    });
-
-    const newApplet = parseAppletEvents(applet);
-
-    let scores = [],
-      maxScores = [];
-    for (let i = 0; i < activity.items.length; i++) {
-      if (!activity.items[i] || !responses[i]) continue;
-
-      let score = getScoreFromResponse(activity.items[i], responses[i]);
-      scores.push(score);
-
-      maxScores.push(getMaxScore(activity.items[i]));
-    }
-
-    const cumulativeScores = activity.compute.reduce((accumulator, itemCompute) => {
-      return {
-        ...accumulator,
-        [itemCompute.variableName.trim().replace(/\s/g, "__")]: evaluateScore(
-          itemCompute.jsExpression,
-          activity.items,
-          scores
-        ),
-      };
-    }, {});
-
-    const cumulativeMaxScores = activity.compute.reduce((accumulator, itemCompute) => {
-      return {
-        ...accumulator,
-        [itemCompute.variableName.trim().replace(/\s/g, "__")]: evaluateScore(
-          itemCompute.jsExpression,
-          activity.items,
-          maxScores
-        ),
-      };
-    }, {});
-
-    const reportMessages = [];
-    let cumActivities = [];
-    activity.messages.forEach(async (msg, i) => {
-      const { jsExpression, message, outputType, nextActivity } = msg;
-
-      const exprArr = jsExpression.split(/[><]/g);
-      const variableName = exprArr[0];
-      const exprValue = parseFloat(exprArr[1].split(" ")[1]);
-      const category = variableName.trim().replace(/\s/g, "__");
-      const expr = parser.parse(category + jsExpression.substr(variableName.length));
-
-      const variableScores = {
-        [category]:
-          outputType == "percentage"
-            ? Math.round(
-                cumulativeMaxScores[category] ? (cumulativeScores[category] * 100) / cumulativeMaxScores[category] : 0
-              )
-            : cumulativeScores[category],
-      };
-
-      if (expr.evaluate(variableScores)) {
-        if (nextActivity) cumActivities.push(nextActivity);
-
-        const compute = activity.compute.find((itemCompute) => itemCompute.variableName.trim() == variableName.trim());
-
-        reportMessages.push({
-          category,
-          message,
-          score: variableScores[category] + (outputType == "percentage" ? "%" : ""),
-          compute,
-          jsExpression: jsExpression.substr(variableName.length),
-          scoreValue: cumulativeScores[category],
-          maxScoreValue: cumulativeMaxScores[category],
-          exprValue: outputType == "percentage" ? (exprValue * cumulativeMaxScores[category]) / 100 : exprValue,
-        });
-      }
-    });
+    let { reportMessages, cumActivities } = evaluateCumulatives(responses, activity)
 
     if (cumulativeActivities && cumulativeActivities[`${activity.id}/nextActivity`]) {
       cumActivities = _.difference(cumActivities, cumulativeActivities[`${activity.id}/nextActivity`]);
@@ -279,7 +203,7 @@ const ActivitySummary = ({ responses, activity, applet, cumulativeActivities, se
           </b>
         </p>
         <p class="text-body-2 mb-4">
-          Your/Your child’s score on the 
+          Your/Your child’s score on the
           ${message.category.replace(/_/g, " ")}
            subscale was <span class="text-danger">${message.scoreValue}</span>.
           ${markdownItInstance.render(message.message)}
@@ -381,6 +305,9 @@ const ActivitySummary = ({ responses, activity, applet, cumulativeActivities, se
           margin: 0;
           right: 0;
           bottom: 0;
+        }
+        img {
+          max-width: 100%;
         }
       </style>
     `;
